+++ packages/dakanji_db_core/lib/extensions/sqlite_vector_extension.dart
@@ -5,7 +5,7 @@ import 'dart:ffi';
 import 'package:sqlite3/sqlite3.dart';
 
 @Native<Int Function(Pointer<Void>, Pointer<Void>, Pointer<Void>)>()
-external int sqlite3_vec_init(
+external int vector_init(
   Pointer<Void> db,
   Pointer<Void> pzErrMsg,
   Pointer<Void> pApi,
@@ -19,7 +19,7 @@ extension LoadVectorExtension on Sqlite3 {
           NativeFunction<
             Int Function(Pointer<Void>, Pointer<Void>, Pointer<Void>)
           >
-        >(sqlite3_vec_init).cast(),
+        >(vector_init).cast(),
       ),
     );
   }