+++ packages/dakanji_db_core/test/extensions/test_extensions_loaded.dart
@@ -1,13 +1,13 @@
 import 'package:test/test.dart';
 import 'package:sqlite3/native_assets.dart';
-import 'package:dakanji_db_core/extensions/sqlite_vec_extension.dart';
+import 'package:dakanji_db_core/extensions/sqlite_vector_extension.dart';
 import 'package:dakanji_db_core/extensions/sqlite_spellfix_extension.dart';
 import 'package:dakanji_db_core/extensions/sqlite_crsqlite_extension.dart';
-
+import 'package:dakanji_db_core/extensions/sqlite_compress_extension.dart';
 
 void main() async {
 
-  sqlite3Native.loadSqliteVecExtension();
+  //sqlite3Native.loadSqliteVecExtension();
   sqlite3Native.loadSqliteSpellfixExtension();
   sqlite3Native.loadSqliteCrsqliteExtension();
 
@@ -28,6 +28,10 @@ void main() async {
   test('Testing SQLite-vec loaded', () async {
     db.execute('create virtual table vec_examples using vec0(sample_embedding float[8]);');
   });
+
+  //test('Testing SQLite-vec loaded', () async {
+  //  db.execute('create virtual table vec_examples using vec0(sample_embedding float[8]);');
+  //});
 
   test('Testing CR-SQLite loaded', () async {
     db.execute('''